[ADD] Automatic portal Account creation for new Registry Task Id:3427361
Modifies motorcycle.registry so that whenever a new registry is created,
verifies if it has an associated sales order with an email so that a
portal account may be created for the owner.
Method checks if an existing user with the given email is already
present (if so, no changes will be done to prevent overwriting existing
permissions) and if not creates a new user with portal access level.

Extended res.users to create a new method for sending signup invitations
using a customized email template used for the aforementioned Account
invite to the portal.

Task Id:3427361 Dev: Odoo Team 4